Внес правки по компоненту логотип, декомпозировал главную страницу - вынес каталог в отдельный компонент
