+++ project/src/components/app/app.tsx
@@ -10,20 +10,12 @@ import PlayerScreen from '../../pages/player-screen/player-screen';
 import PrivateRoute from '../private-route/private-route';
 import Tabs from '../tabs/tabs';
 import { useAppSelector } from '../../hooks';
-import Spinner from '../spinner/spinner';
 import HistoryRouter from '../history-route/history-route';
 import browserHistory from '../../browser-history';
 
 function App(): JSX.Element {
   const films = useAppSelector((state) => state.films);
   const reviews = useAppSelector((state) => state.reviews);
-  const isFilmsLoading = useAppSelector((state) => state.isFilmsLoading);
-
-  if (isFilmsLoading) {
-    return (
-      <Spinner />
-    );
-  }
 
   return (
     <HistoryRouter history={browserHistory}>
