+++ project/src/pages/login-screen/login-screen.tsx
@@ -64,7 +64,7 @@ function LoginScreen(): JSX.Element {
 
       <footer className="page-footer">
         <div className="logo">
-          <Logo />
+          <Logo isFooter/>
         </div>
 
         <div className="copyright">
