+++ project/src/pages/my-list-screen/my-list-screen.tsx
@@ -37,7 +37,7 @@ function MyListScreen({ films }: Props): JSX.Element {
 
       <footer className="page-footer">
         <div className="logo">
-          <Logo />
+          <Logo isFooter/>
         </div>
 
         <div className="copyright">
